requirements.txt
fastapi
paho-mqtt
pyjwt
python-dotenv
sqlmodel
uvicorn
mysql-connector-python
dependency-injector